YSD/requirements.txt
a-pytube-fork-for-spotdl-users==9.6.4
absl-py==0.11.0
aiofiles==0.6.0
aiohttp==3.7.4.post0
appdirs==1.4.4
APScheduler==3.6.3
astunparse==1.6.3
async-timeout==3.0.1
attrs==21.2.0
chardet==3.0.4
colorama==0.4.4
coloredlogs==15.0
commonmark==0.9.1
configparser==5.0.2
crayons==0.4.0
cryptography==3.2.1
flatbuffers==1.12
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2021.7.5
httpcore==0.9.1
httpx==0.13.3
humanfriendly==9.1
hyperframe==5.2.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==1.0.1
kaitaistruct==0.9
langdetect==1.0.8
lxml==4.6.3
Markdown==3.3.4
MarkupSafe==1.1.1
MouseInfo==0.1.3
multidict==5.1.0
multipart==0.2.4
mutagen==1.45.1
py-deezer==1.1.4.post1
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.52
pycparser==2.20
pydub==0.24.1
PyGetWindow==0.0.9
Pygments==2.9.0
PyInquirer==1.0.3
PyMsgBox==1.0.9
pyperclip==1.8.2
pyreadline==2.1
PyRect==0.1.4
PyScreeze==0.1.27
pyTelegramBotAPI==3.7.4
python-telegram-bot==13.13
pytube==10.4.1
PyTweening==1.0.3
pytz==2020.1
PyYAML==5.4.1
psycopg2-binary==2.9.3
rapidfuzz==0.12.5
regex==2020.11.13
requests==2.25.0
requests-oauthlib==1.3.0
rfc3986==1.4.0
rsa==4.7.2
ShazamAPI==0.0.2
six==1.15.0
sniffio==1.2.0
soupsieve==2.2
spotipy==2.16.1
svgwrite==1.4.1
tzlocal==2.1
tqdm==4.62.3
unicode-slugify==0.1.3
Unidecode==1.2.0
urllib3==1.26.2
wcwidth==0.2.5
webdriver-manager==3.4.0
websocket-client==0.59.0
yarl==1.6.3
youtube-dl==2021.12.17
yt-dlp==2022.7.18
ytmusicapi==0.14.2
